Projectiles, Characters Tweaks, and Fixes
ShootyMcShootFace is now a subclass of CatCharacter
Projectile now detects hitting a CatCharacter.
Projectiles remember who shot them.
Projectiles do damage if the shooter and the cat hit are not friends.
Projectile now has a Damage property.
Fixed a few issues with aiming in the CatCharacter class.
Slowed down the climb speed.